Revert "[3기 김종완] TodoList for team (#16)"
This reverts commit a927427996bf837adad0e8405e7a0873ba152a2d.
