+++ README.md
@@ -110,7 +110,7 @@
 }
 ```
 
-### 팀원별 UserListContainer 불러오기
+### 팀원별 TodoList 불러오기
 
 | method | uri |
 |---|---|
