+++ index.html
@@ -1,22 +1,52 @@
 <!DOCTYPE html>
 <html>
-<head>
+  <head>
     <title>Mission 3 Todo List For Team</title>
-    <meta charset="utf-8"/>
-    <link href="./css/style.css" rel="stylesheet"/>
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet"/>
-</head>
-<body>
-
-<div id="app">
-    <h1 id="user-title" data-username="eastjun">
-        <span><strong>김종완</strong>'s Todo Lists</span>
-    </h1>
-    <div class="team-list-container"></div>
-    <section class="input-container">
-    </section>
-</div>
-
-<script src="./js/index.js" type="module"></script>
-</body>
+    <meta charset="utf-8" />
+    <link href="./css/style.css" rel="stylesheet" />
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet" />
+  </head>
+  <body>
+    <div id="app">
+      <h1 id="user-title" data-username="eastjun">
+        <span><strong>Team</strong>'s Todo Lists</span>
+      </h1>
+      <div class="team-list-container">
+        <div class="team-card-container">
+          <a href="/kanban.html" class="card">
+            <div class="card-title">
+              Black Coffee team
+            </div>
+          </a>
+        </div>
+        <div class="team-card-container">
+          <a href="/kanban.html" class="card">
+            <div class="card-title">
+              Black Coffee team
+            </div>
+          </a>
+        </div>
+        <div class="team-card-container">
+          <a href="/kanban.html" class="card">
+            <div class="card-title">
+              Black Coffee team
+            </div>
+          </a>
+        </div>
+        <div class="team-card-container">
+          <a href="/kanban.html" class="card">
+            <div class="card-title">
+              Black Coffee team
+            </div>
+          </a>
+        </div>
+        <div class="add-team-button-container">
+          <button id="add-team-button" class="ripple">
+            <span class="material-icons">add</span>
+          </button>
+        </div>
+      </div>
+    </div>
+    <script src="./js/Team.js"></script>
+  </body>
 </html>
