+++ js/Team.js
@@ -1,0 +1,4 @@
+const $addTeamButton = document.querySelector('#add-team-button')
+$addTeamButton.addEventListener('click', () => {
+  const result = prompt('팀 이름을 입력해주세요')
+})
