--- js/constants/constants.js
@@ -1,34 +1,0 @@
-export const path = {
-    INDEX_HTML: 'index.html',
-    KANBAN_HTML: 'kanban.html'
-
-}
-export const baseUrl = 'https://js-todo-list-9ca3a.df.r.appspot.com'
-
-export const method = {
-    GET : "GET",
-    POST : "POST",
-    PUT : "PUT",
-    DELETE : "DELETE",
-}
-export const priority = {
-    FIRST : "FIRST",
-    SECOND : "SECOND",
-    NONE : "NONE",
-}
-export const keyboardKey = {
-    Enter : 'Enter',
-    Esc : 'Escape'
-}
-export const eventType = {
-    CLICK : 'click',
-    DOUBLE_CLICK : 'dblclick',
-    CHANGE : 'change',
-    KEY_UP : 'keyup',
-}
-export const filter = {
-    ALL : 'all',
-    PRIORITY : 'priority',
-    ACTIVE : 'active',
-    COMPLETED : 'completed',
-}
