--- js/store/index.js
@@ -1,11 +1,0 @@
-import actions from './actions.js';
-import mutations from "./mutations.js";
-import state from "./state.js";
-import Store from "./store.js";
-
-
-export default new Store({
-    actions,
-    mutations,
-    state
-})
