--- js/store/state.js
@@ -1,14 +1,0 @@
-export default{
-    /*
-    * team
-    *
-    * id : string
-    * name : string
-    * members : [ ...]
-    * */
-    team : [],
-    selectedTeam : [],
-
-
-
-}
